feat(order): 订单管理功能完善
- 修改分页查询DTO中page和pageSize字段类型为Long
- 新增OrderCancelDTO、OrderConfirmDTO用于订单操作
- 移除订单常量中的退款状态定义
- 新增管理员订单控制器实现订单取消、统计、完成、拒单、接单等功能
- 用户订单控制器增加历史订单查询、订单详情、取消订单、再来一单功能
- 重构订单查询DTO和VO数据传输对象
- 新增订单搜索DTO和VO支持条件查询
- 新增订单统计VO用于订单状态统计
- 完善订单服务接口及实现类提供完整的订单生命周期管理
- 在工具文档中补充Stream流分组和连接操作说明

+++ sky-take-out/sky-common/src/main/java/com/hongs/skycommon/pojo/dto/DishPageQueryDTO.java
@@ -19,8 +19,8 @@ public class DishPageQueryDTO implements Serializable {
     private Integer status;
 
     @Schema(description = "页码")
-    private int page;
+    private Long page;
 
     @Schema(description = "每页记录数")
-    private int pageSize;
+    private Long pageSize;
 }
